refactor: create "simple" wallet data object (#297)
