+++ app/Data/SimpleWalletData.php
@@ -2,17 +2,20 @@
 
 declare(strict_types=1);
 
-namespace App\Data\Nfts;
+namespace App\Data;
 
+use App\Data\Wallet\WalletAvatarData;
 use App\Models\Wallet;
 use Spatie\LaravelData\Data;
 use Spatie\TypeScriptTransformer\Attributes\TypeScript;
 
 #[TypeScript]
-class NftWalletData extends Data
+class SimpleWalletData extends Data
 {
     public function __construct(
         public string $address,
+        public ?string $domain,
+        public WalletAvatarData $avatar,
     ) {
     }
 
@@ -20,6 +23,8 @@ class NftWalletData extends Data
     {
         return new self(
             address: $wallet->address,
+            domain: $wallet->details?->domain,
+            avatar: WalletAvatarData::fromModel($wallet),
         );
     }
 }
