+++ resources/js/Tests/Factories/Nfts/NftFactory.ts
@@ -2,7 +2,7 @@ import { faker } from "@faker-js/faker";
 import ModelFactory from "@/Tests/Factories/ModelFactory";
 import NFTCollectionFactory from "@/Tests/Factories/Nfts/NFTCollectionFactory";
 import NftImagesDataFactory from "@/Tests/Factories/Nfts/NftImagesDataFactory";
-import NftWalletFactory from "@/Tests/Factories/Nfts/NftWalletFactory";
+import SimpleWalletDataFactory from "@/Tests/Factories/SimpleWalletDataFactory";
 
 export default class NftFactory extends ModelFactory<App.Data.Nfts.NftData> {
     protected factory(): App.Data.Nfts.NftData {
@@ -13,7 +13,7 @@ export default class NftFactory extends ModelFactory<App.Data.Nfts.NftData> {
             tokenNumber: faker.datatype.number({ min: 1, max: 100000 }).toString(),
             collection: new NFTCollectionFactory().withImage().create(),
             images: new NftImagesDataFactory().create(),
-            wallet: this.optional(new NftWalletFactory().create()),
+            wallet: this.optional(new SimpleWalletDataFactory().create()),
             lastViewedAt: null,
             lastActivityFetchedAt: null,
         };
@@ -21,7 +21,7 @@ export default class NftFactory extends ModelFactory<App.Data.Nfts.NftData> {
 
     public withWallet(): this {
         return this.state(() => ({
-            wallet: new NftWalletFactory().create(),
+            wallet: new SimpleWalletDataFactory().create(),
         }));
     }
